+++ tikloo/tk_default_draw.h
@@ -4,7 +4,7 @@
 #include"default_draw/draw_flatDial.h"
 
 //dial
-void cairo_code_draw_flatDial_render(cairo_t *cr, uint16_t w, uint16_t h, void* valp);
+void cairo_code_draw_flatDial_render(cairo_t *cr, float w, float h, void* valp);
